Create Examination
Two friends, Dragon and Sloth, are writing a computer science examination series. There are three subjects in this series:DSA, TOC,DM.You know the individual scores of both Dragon and Sloth in all 3 subjects. You have to determine who got a better rank.
The rank is decided as follows:
The person with a bigger total score gets a better rankIf the total scores are tied, the person who scored higher in DSA gets a better rank If the total score and the DSA score are tied, the person who scored higher in TOC gets a better rank
If everything is tied, they get the same rank.