Create Candies
You are given an array A of size 2𝑁. Find out whether it is possible to split A into two arrays, each of length 𝑁, such that both arrays consist of distinct elements. Both arrays can have distinct elements only if no element in the original array is repeated more than twice.