Create  Periodic String
Given a string ‘S’ find whether the given string is periodic or not.
A string is said to be periodic if it repeats itself after a certain number of characters and the period of repetition is less than the size of the string.For example: Let ‘S’ be “ababab” we can clearly see that this string is periodic as ‘ab’ is repeated 3 times to make the string ‘S’.
Sample Input 1:
3
xxxxxx
aabbaaabba
abcba
Sample Output 1:
True
True
False      
Explanation of sample input 1 :
Test Case 1:
In the first test case, we can clearly see that the string has a repeating string ‘x’ 6 times. So we return tru
Test Case 2:
In the second test case, we can see that the string ‘aabba’ repeats twice to form the given string. Hence we return true,
Test Case 3:
In the third test case, we can see that there is no string which repeats itself to form the given string, hence we return false.